Refactor BaseText component as Base component
- Issue: https://git.chen.so/tina-next-ts/i/bdea9755b343dfa96250c621694cc6c17b14c846
